Move error handling out of __init__.py
